Fix underscored functions in abi.rs
