+++ src/evm/abi.rs
@@ -151,15 +151,6 @@ impl AbiForContract {
         self.contract.function(name)
     }
 
-    pub fn _generate_calldata_for_function(
-        &self,
-        func_name: &str,
-        args: &[ethabi::Token],
-    ) -> Result<Vec<u8>, ethabi::Error> {
-        let this_function = self.contract.function(func_name)?;
-        Ok(this_function.encode_input(args).unwrap())
-    }
-
     pub fn call_function(
         &self,
         sender_addr: Uint256,
@@ -180,114 +171,6 @@ impl AbiForContract {
             payment,
             &calldata,
             false,
-        );
-
-        let num_logs_before = machine.runtime_env.get_all_receipt_logs().len();
-        let num_sends_before = machine.runtime_env.get_all_sends().len();
-        let _arbgas_used = if debug {
-            machine.debug(None)
-        } else {
-            machine.run(None)
-        };
-        let logs = machine.runtime_env.get_all_receipt_logs();
-        let sends = machine.runtime_env.get_all_sends();
-        Ok((
-            logs[num_logs_before..].to_vec(),
-            sends[num_sends_before..].to_vec(),
-        ))
-    }
-
-    pub fn _send_retryable_tx(
-        &self,
-        sender: Uint256,
-        func_name: &str,
-        args: &[ethabi::Token],
-        machine: &mut Machine,
-        callvalue: Uint256,
-        deposit: Uint256,
-        max_submission_cost: Uint256,
-        credit_back_address: Option<Uint256>,
-        beneficiary: Option<Uint256>,
-        max_gas_immed: Option<Uint256>,
-        gas_price_immed: Option<Uint256>,
-    ) -> Result<(Uint256, Uint256, Option<Uint256>), ethabi::Error> {
-        let this_function = self.contract.function(func_name)?;
-        let calldata = this_function.encode_input(args).unwrap();
-
-        let (submit_txid, retryable_txid, maybe_redeemid) =
-            machine.runtime_env._insert_retryable_tx_message(
-                sender.clone(),
-                self.address.clone(),
-                callvalue,
-                deposit,
-                max_submission_cost,
-                credit_back_address.unwrap_or(sender.clone()),
-                beneficiary.unwrap_or(sender),
-                max_gas_immed.unwrap_or(Uint256::zero()),
-                gas_price_immed.unwrap_or(Uint256::zero()),
-                &calldata,
-            );
-
-        Ok((submit_txid, retryable_txid, maybe_redeemid))
-    }
-
-    pub fn _call_function_from_contract(
-        &self,
-        sender_addr: Uint256,
-        func_name: &str,
-        args: &[ethabi::Token],
-        machine: &mut Machine,
-        payment: Uint256,
-        debug: bool,
-    ) -> Result<(Vec<ArbosReceipt>, Vec<Vec<u8>>), ethabi::Error> {
-        let this_function = self.contract.function(func_name)?;
-        let calldata = this_function.encode_input(args).unwrap();
-
-        machine.runtime_env.insert_tx_message_from_contract(
-            sender_addr,
-            Uint256::from_usize(100_000_000),
-            None,
-            self.address.clone(),
-            payment,
-            &calldata,
-            false,
-        );
-
-        let num_logs_before = machine.runtime_env.get_all_receipt_logs().len();
-        let num_sends_before = machine.runtime_env.get_all_sends().len();
-        let _arbgas_used = if debug {
-            machine.debug(None)
-        } else {
-            machine.run(None)
-        };
-        let logs = machine.runtime_env.get_all_receipt_logs();
-        let sends = machine.runtime_env.get_all_sends();
-        Ok((
-            logs[num_logs_before..].to_vec(),
-            sends[num_sends_before..].to_vec(),
-        ))
-    }
-
-    pub fn _call_function_with_deposit(
-        &self,
-        sender_addr: Uint256,
-        func_name: &str,
-        args: &[ethabi::Token],
-        machine: &mut Machine,
-        payment: Uint256,
-        debug: bool,
-    ) -> Result<(Vec<ArbosReceipt>, Vec<Vec<u8>>), ethabi::Error> {
-        let this_function = self.contract.function(func_name)?;
-        let calldata = this_function.encode_input(args).unwrap();
-
-        machine.runtime_env.insert_tx_message(
-            sender_addr,
-            Uint256::from_usize(100_000_000),
-            None,
-            self.address.clone(),
-            payment,
-            &calldata,
-            true,
         );
 
         let num_logs_before = machine.runtime_env.get_all_receipt_logs().len();
@@ -373,33 +256,6 @@ impl AbiForContract {
         Ok((Uint256::from_bytes(&tx_id_bytes)))
     }
 
-    pub fn _add_function_call_to_compressed_batch(
-        &self,
-        batch: &mut Vec<u8>,
-        func_name: &str,
-        args: &[ethabi::Token],
-        machine: &mut Machine,
-        payment: Uint256,
-        wallet: &Wallet,
-    ) -> Result<Uint256, ethabi::Error> {
-        let this_function = self.contract.function(func_name)?;
-        let calldata = this_function.encode_input(args).unwrap();
-
-        let tx_id_bytes = machine
-            .runtime_env
-            ._append_compressed_and_signed_tx_message_to_batch(
-                batch,
-                Uint256::from_usize(100_000_000),
-                None,
-                self.address.clone(),
-                payment,
-                calldata,
-                &wallet,
-            );
-
-        Ok((Uint256::from_bytes(&tx_id_bytes)))
-    }
-
     pub fn short_signature_for_function(&self, func_name: &str) -> Result<[u8; 4], ethabi::Error> {
         let long_sig = self.contract.function(func_name)?.signature();
         let short_sig = long_sig.split(":").collect::<Vec<&str>>()[0];
@@ -452,6 +308,153 @@ impl AbiForContract {
     }
 }
 
+#[cfg(test)]
+impl AbiForContract {
+    pub fn generate_calldata_for_function(
+        &self,
+        func_name: &str,
+        args: &[ethabi::Token],
+    ) -> Result<Vec<u8>, ethabi::Error> {
+        let this_function = self.contract.function(func_name)?;
+        Ok(this_function.encode_input(args).unwrap())
+    }
+
+    pub fn send_retryable_tx(
+        &self,
+        sender: Uint256,
+        func_name: &str,
+        args: &[ethabi::Token],
+        machine: &mut Machine,
+        callvalue: Uint256,
+        deposit: Uint256,
+        max_submission_cost: Uint256,
+        credit_back_address: Option<Uint256>,
+        beneficiary: Option<Uint256>,
+        max_gas_immed: Option<Uint256>,
+        gas_price_immed: Option<Uint256>,
+    ) -> Result<(Uint256, Uint256, Option<Uint256>), ethabi::Error> {
+        let this_function = self.contract.function(func_name)?;
+        let calldata = this_function.encode_input(args).unwrap();
+
+        let (submit_txid, retryable_txid, maybe_redeemid) =
+            machine.runtime_env._insert_retryable_tx_message(
+                sender.clone(),
+                self.address.clone(),
+                callvalue,
+                deposit,
+                max_submission_cost,
+                credit_back_address.unwrap_or(sender.clone()),
+                beneficiary.unwrap_or(sender),
+                max_gas_immed.unwrap_or(Uint256::zero()),
+                gas_price_immed.unwrap_or(Uint256::zero()),
+                &calldata,
+            );
+
+        Ok((submit_txid, retryable_txid, maybe_redeemid))
+    }
+
+    pub fn call_function_from_contract(
+        &self,
+        sender_addr: Uint256,
+        func_name: &str,
+        args: &[ethabi::Token],
+        machine: &mut Machine,
+        payment: Uint256,
+        debug: bool,
+    ) -> Result<(Vec<ArbosReceipt>, Vec<Vec<u8>>), ethabi::Error> {
+        let this_function = self.contract.function(func_name)?;
+        let calldata = this_function.encode_input(args).unwrap();
+
+        machine.runtime_env.insert_tx_message_from_contract(
+            sender_addr,
+            Uint256::from_usize(100_000_000),
+            None,
+            self.address.clone(),
+            payment,
+            &calldata,
+            false,
+        );
+
+        let num_logs_before = machine.runtime_env.get_all_receipt_logs().len();
+        let num_sends_before = machine.runtime_env.get_all_sends().len();
+        let _arbgas_used = if debug {
+            machine.debug(None)
+        } else {
+            machine.run(None)
+        };
+        let logs = machine.runtime_env.get_all_receipt_logs();
+        let sends = machine.runtime_env.get_all_sends();
+        Ok((
+            logs[num_logs_before..].to_vec(),
+            sends[num_sends_before..].to_vec(),
+        ))
+    }
+
+    pub fn call_function_with_deposit(
+        &self,
+        sender_addr: Uint256,
+        func_name: &str,
+        args: &[ethabi::Token],
+        machine: &mut Machine,
+        payment: Uint256,
+        debug: bool,
+    ) -> Result<(Vec<ArbosReceipt>, Vec<Vec<u8>>), ethabi::Error> {
+        let this_function = self.contract.function(func_name)?;
+        let calldata = this_function.encode_input(args).unwrap();
+
+        machine.runtime_env.insert_tx_message(
+            sender_addr,
+            Uint256::from_usize(100_000_000),
+            None,
+            self.address.clone(),
+            payment,
+            &calldata,
+            true,
+        );
+
+        let num_logs_before = machine.runtime_env.get_all_receipt_logs().len();
+        let num_sends_before = machine.runtime_env.get_all_sends().len();
+        let _arbgas_used = if debug {
+            machine.debug(None)
+        } else {
+            machine.run(None)
+        };
+        let logs = machine.runtime_env.get_all_receipt_logs();
+        let sends = machine.runtime_env.get_all_sends();
+        Ok((
+            logs[num_logs_before..].to_vec(),
+            sends[num_sends_before..].to_vec(),
+        ))
+    }
+
+    pub fn add_function_call_to_compressed_batch(
+        &self,
+        batch: &mut Vec<u8>,
+        func_name: &str,
+        args: &[ethabi::Token],
+        machine: &mut Machine,
+        payment: Uint256,
+        wallet: &Wallet,
+    ) -> Result<Uint256, ethabi::Error> {
+        let this_function = self.contract.function(func_name)?;
+        let calldata = this_function.encode_input(args).unwrap();
+
+        let tx_id_bytes = machine
+            .runtime_env
+            ._append_compressed_and_signed_tx_message_to_batch(
+                batch,
+                Uint256::from_usize(100_000_000),
+                None,
+                self.address.clone(),
+                payment,
+                calldata,
+                &wallet,
+            );
+
+        Ok((Uint256::from_bytes(&tx_id_bytes)))
+    }
+}
+
 pub fn deploy_add(machine: &mut Machine) -> Result<AbiForContract, ethabi::Error> {
     AbiForContract::new_deployed(
         machine,
@@ -490,7 +493,7 @@ impl<'a> ArbSys<'a> {
         }
     }
 
-    pub fn _arbos_version(&self, machine: &mut Machine) -> Result<Uint256, ethabi::Error> {
+    pub fn arbos_version(&self, machine: &mut Machine) -> Result<Uint256, ethabi::Error> {
         let (receipts, _sends) = self.contract_abi.call_function_compressed(
             self.my_address.clone(),
             "arbOSVersion",
@@ -514,35 +517,6 @@ impl<'a> ArbSys<'a> {
                 ethabi::Token::Uint(ui) => Ok(Uint256::from_u256(&ui)),
                 _ => panic!(),
             }
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _withdraw_eth(
-        &self,
-        machine: &mut Machine,
-        payee_addr: Uint256,
-        amount: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function_compressed(
-            self.my_address.clone(),
-            "withdrawEth",
-            &[ethabi::Token::Address(ethabi::Address::from(
-                payee_addr.to_h160(),
-            ))],
-            machine,
-            amount,
-            self._wallet,
-            self.debug,
-        )?;
-
-        if (receipts.len() != 1) {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
         } else {
             Err(ethabi::Error::from("reverted"))
         }
